service: Signal when services are moved
When calling MoveBefore or MoveAfter on a service, no signal is emitted
by ConnMan. This makes it impossible for a D-Bus client to maintain an
ordered list of services in sync with ConnMan.

This patch schedules the emission of the signal ServicesChanged from the
function move_service(), so the new list will get published over D-Bus.