test: Add tests and test tool for IPv6 parts of iptables.c.
This commit adds tests for IPv6 enabled iptables. The tests are
identical to the existing iptables tests, except IPv6 "nat" table rules
are not tested as IPv6 NAT is not enabled.

Also a test tool for IPv6 iptables (ip6tables-test) has been added,
which is a clone of iptables-test. iptables-test.c has been modified to
support the changes in iptables.c.

Added ip6tables-save program to configure.ac and use of it in
Makefile.am for the updated iptables-unit test.